refactor(frontend): migrate auth to Zustand store and delete legacy AuthContext (FR-007 US2.6)
