+++ frontend/src/__tests__/icon-font/preservation.property.test.tsx
@@ -57,11 +57,10 @@ vi.mock('next/link', () => ({
     React.createElement('a', { href, ...props }, children),
 }));
 
-// Mock AuthContext — Navbar uses useAuth
-vi.mock('@/contexts/AuthContext', () => ({
-  AuthProvider: ({ children }: { children: React.ReactNode }) =>
-    React.createElement(React.Fragment, null, children),
-  useAuth: () => ({ isAuthenticated: false, user: null, logout: vi.fn() }),
+// Mock shared store — Navbar uses useAuth
+vi.mock('@/shared/store', () => ({
+  useAuth: () => ({ isAuthenticated: false, user: null, status: 'unauthenticated', logout: vi.fn() }),
+  useTheme: () => ({ mode: 'dark', setMode: vi.fn() }),
 }));
 
 // Import components AFTER mocks are set up
