+++ frontend/src/app/auth/forbidden/page.tsx
@@ -4,7 +4,7 @@
 // Requirements: 13.2, 13.6
 
 import Link from "next/link";
-import { useAuth } from "@/contexts/AuthContext";
+import { useAuth } from "@/shared/store";
 import { Icon } from "@/shared/ui/Icon";
 
 const ROLE_LABELS: Record<string, string> = {
